Final Draft Audio EQ Cookbook for HP15c CE
This is a fully working version. MUST be loaded and run under 15.2 mode. Size 442 bytes (plus an easter egg: Prog 0 converts to dB from linear and Prog 1 converts to linear from dB.  Easter egg is only available in the .bin file.
Note: The compiler has a bug when implementing RESULT instruction.  It uses the wrong opcode. This does not affect the .hex and the .bin, only the versions that use opcodes like .15c and .jrn files.
TODO: Documentation for users and release.