DFT program did not for MATRIX 1
DFT program did not do a MATRIX 1 at the end of Prog B and will cause user an get an Error 3 if they try to RCL Matrix A or B right after running.  Setting MATRIX 1 will avoid confusion to user and save time.  Also, delete the .hex files since we moved on to using the PH15C_Unterface.xls tool.